rag-lambda/requirements.txt
# Todo: Automatically package these dependencies
langchain-openai
langchain-pinecone
pinecone
python-dotenv
urllib3==1.26.7